requirements.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.1
certifi==2018.4.16
chardet==3.0.4
Django==2.0.4
django-filter==1.1.0
djangorestframework==3.8.2
gnureadline==6.3.8
idna==2.7
kombu==4.2.1
Markdown==2.6.11
numpy==1.14.2
pandas==0.22.0
pymongo==3.6.1
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
scikit-learn==0.19.1
scipy==1.0.1
six==1.11.0
sklearn==0.0
urllib3==1.23
vine==1.1.4
